+++ packages/modules-sdk/src/loaders/__mocks__/@plugins/default.ts
@@ -1,0 +1,5 @@
+const service = class TestService {}
+
+export default {
+  services: [service],
+}