+++ packages/stock-location/src/services/stock-location.ts
@@ -12,12 +12,12 @@ import {
 } from "@medusajs/types"
 import {
   InjectEntityManager,
+  isDefined,
   MedusaContext,
   MedusaError,
-  isDefined,
   setMetadata,
 } from "@medusajs/utils"
-import { EntityManager } from "typeorm"
+import { EntityManager, FindOptionsWhere, ILike } from "typeorm"
 import { joinerConfig } from "../joiner-config"
 import { StockLocation, StockLocationAddress } from "../models"
 import { buildQuery } from "../utils/build-query"
@@ -64,13 +64,10 @@ export default class StockLocationService {
   async list(
     selector: FilterableStockLocationProps = {},
     config: FindConfig<StockLocation> = { relations: [], skip: 0, take: 10 },
-    context: SharedContext = {}
+    @MedusaContext() context: SharedContext = {}
   ): Promise<StockLocation[]> {
-    const manager = context.transactionManager ?? this.manager_
-    const locationRepo = manager.getRepository(StockLocation)
-
-    const query = buildQuery(selector, config)
-    return await locationRepo.find(query)
+    const [locations] = await this.listAndCount(selector, config, context)
+    return locations
   }
 
   /**
@@ -83,12 +80,31 @@ export default class StockLocationService {
   async listAndCount(
     selector: FilterableStockLocationProps = {},
     config: FindConfig<StockLocation> = { relations: [], skip: 0, take: 10 },
-    context: SharedContext = {}
+    @MedusaContext() context: SharedContext = {}
   ): Promise<[StockLocation[], number]> {
     const manager = context.transactionManager ?? this.manager_
     const locationRepo = manager.getRepository(StockLocation)
+    let q
+    if (selector.q) {
+      q = selector.q
+      delete selector.q
+    }
 
     const query = buildQuery(selector, config)
+
+    if (q) {
+      const where = query.where as FindOptionsWhere<StockLocation>
+
+      delete where.name
+
+      query.where = [
+        {
+          ...where,
+          name: ILike(`%${q}%`),
+        },
+      ]
+    }
+
     return await locationRepo.findAndCount(query)
   }
 
@@ -103,7 +119,7 @@ export default class StockLocationService {
   async retrieve(
     stockLocationId: string,
     config: FindConfig<StockLocation> = {},
-    context: SharedContext = {}
+    @MedusaContext() context: SharedContext = {}
   ): Promise<StockLocation> {
     if (!isDefined(stockLocationId)) {
       throw new MedusaError(