+++ packages/types/src/customer/service.ts
@@ -1,14 +1,26 @@
 import { FindConfig } from "../common"
+import { RestoreReturn, SoftDeleteReturn } from "../dal"
 import { IModuleService } from "../modules-sdk"
 import { Context } from "../shared-context"
 import {
   CustomerDTO,
   CustomerGroupDTO,
+  CustomerGroupCustomerDTO,
+  FilterableCustomerGroupCustomerProps,
   FilterableCustomerProps,
   FilterableCustomerGroupProps,
   GroupCustomerPair,
+  FilterableCustomerAddressProps,
+  CustomerAddressDTO,
 } from "./common"
-import { CreateCustomerDTO, CreateCustomerGroupDTO } from "./mutations"
+import {
+  CreateCustomerAddressDTO,
+  CreateCustomerDTO,
+  CreateCustomerGroupDTO,
+  CustomerGroupUpdatableFields,
+  CustomerUpdatableFields,
+  UpdateCustomerAddressDTO,
+} from "./mutations"
 
 export interface ICustomerModuleService extends IModuleService {
   retrieve(
@@ -21,22 +33,21 @@ export interface ICustomerModuleService extends IModuleService {
     data: CreateCustomerDTO[],
     sharedContext?: Context
   ): Promise<CustomerDTO[]>
-
   create(data: CreateCustomerDTO, sharedContext?: Context): Promise<CustomerDTO>
 
   update(
     customerId: string,
-    data: Partial<CreateCustomerDTO>,
+    data: CustomerUpdatableFields,
     sharedContext?: Context
   ): Promise<CustomerDTO>
   update(
     customerIds: string[],
-    data: Partial<CreateCustomerDTO>,
+    data: CustomerUpdatableFields,
     sharedContext?: Context
   ): Promise<CustomerDTO[]>
   update(
     selector: FilterableCustomerProps,
-    data: Partial<CreateCustomerDTO>,
+    data: CustomerUpdatableFields,
     sharedContext?: Context
   ): Promise<CustomerDTO[]>
 
@@ -57,21 +68,37 @@ export interface ICustomerModuleService extends IModuleService {
     sharedContext?: Context
   ): Promise<CustomerGroupDTO>
 
+  retrieveCustomerGroup(
+    groupId: string,
+    config?: FindConfig<CustomerGroupDTO>,
+    sharedContext?: Context
+  ): Promise<CustomerGroupDTO>
+
   updateCustomerGroup(
     groupId: string,
-    data: Partial<CreateCustomerGroupDTO>,
+    data: CustomerGroupUpdatableFields,
     sharedContext?: Context
   ): Promise<CustomerGroupDTO>
   updateCustomerGroup(
     groupIds: string[],
-    data: Partial<CreateCustomerGroupDTO>,
+    data: CustomerGroupUpdatableFields,
     sharedContext?: Context
   ): Promise<CustomerGroupDTO[]>
   updateCustomerGroup(
     selector: FilterableCustomerGroupProps,
-    data: Partial<CreateCustomerGroupDTO>,
+    data: CustomerGroupUpdatableFields,
     sharedContext?: Context
   ): Promise<CustomerGroupDTO[]>
+
+  deleteCustomerGroup(groupId: string, sharedContext?: Context): Promise<void>
+  deleteCustomerGroup(
+    groupIds: string[],
+    sharedContext?: Context
+  ): Promise<void>
+  deleteCustomerGroup(
+    selector: FilterableCustomerGroupProps,
+    sharedContext?: Context
+  ): Promise<void>
 
   addCustomerToGroup(
     groupCustomerPair: GroupCustomerPair,
@@ -84,13 +111,63 @@ export interface ICustomerModuleService extends IModuleService {
   ): Promise<{ id: string }[]>
 
   removeCustomerFromGroup(
-    groupCustomerPair: { customer_id: string; customer_group_id: string },
+    groupCustomerPair: GroupCustomerPair,
     sharedContext?: Context
   ): Promise<void>
   removeCustomerFromGroup(
-    groupCustomerPairs: { customer_id: string; customer_group_id: string }[],
+    groupCustomerPairs: GroupCustomerPair[],
     sharedContext?: Context
   ): Promise<void>
+
+  addAddresses(
+    addresses: CreateCustomerAddressDTO[],
+    sharedContext?: Context
+  ): Promise<CustomerAddressDTO[]>
+  addAddresses(
+    address: CreateCustomerAddressDTO,
+    sharedContext?: Context
+  ): Promise<CustomerAddressDTO>
+
+  updateAddress(
+    addressId: string,
+    data: UpdateCustomerAddressDTO,
+    sharedContext?: Context
+  ): Promise<CustomerAddressDTO>
+  updateAddress(
+    addressIds: string[],
+    data: UpdateCustomerAddressDTO,
+    sharedContext?: Context
+  ): Promise<CustomerAddressDTO[]>
+  updateAddress(
+    selector: FilterableCustomerAddressProps,
+    data: UpdateCustomerAddressDTO,
+    sharedContext?: Context
+  ): Promise<CustomerAddressDTO[]>
+
+  deleteAddress(addressId: string, sharedContext?: Context): Promise<void>
+  deleteAddress(addressIds: string[], sharedContext?: Context): Promise<void>
+  deleteAddress(
+    selector: FilterableCustomerAddressProps,
+    sharedContext?: Context
+  ): Promise<void>
+
+  listAddresses(
+    filters?: FilterableCustomerAddressProps,
+    config?: FindConfig<CustomerAddressDTO>,
+    sharedContext?: Context
+  ): Promise<CustomerAddressDTO[]>
+
+  listAndCountAddresses(
+    filters?: FilterableCustomerAddressProps,
+    config?: FindConfig<CustomerAddressDTO>,
+    sharedContext?: Context
+  ): Promise<[CustomerAddressDTO[], number]>
+
+  listCustomerGroupRelations(
+    filters?: FilterableCustomerGroupCustomerProps,
+    config?: FindConfig<CustomerGroupCustomerDTO>,
+    sharedContext?: Context
+  ): Promise<CustomerGroupCustomerDTO[]>
 
   list(
     filters?: FilterableCustomerProps,
@@ -115,4 +192,28 @@ export interface ICustomerModuleService extends IModuleService {
     config?: FindConfig<CustomerGroupDTO>,
     sharedContext?: Context
   ): Promise<[CustomerGroupDTO[], number]>
+
+  softDelete<TReturnableLinkableKeys extends string = string>(
+    customerIds: string[],
+    config?: SoftDeleteReturn<TReturnableLinkableKeys>,
+    sharedContext?: Context
+  ): Promise<Record<TReturnableLinkableKeys, string[]> | void>
+
+  restore<TReturnableLinkableKeys extends string = string>(
+    customerIds: string[],
+    config?: RestoreReturn<TReturnableLinkableKeys>,
+    sharedContext?: Context
+  ): Promise<Record<TReturnableLinkableKeys, string[]> | void>
+
+  softDeleteCustomerGroup<TReturnableLinkableKeys extends string = string>(
+    groupIds: string[],
+    config?: SoftDeleteReturn<TReturnableLinkableKeys>,
+    sharedContext?: Context
+  ): Promise<Record<TReturnableLinkableKeys, string[]> | void>
+
+  restoreCustomerGroup<TReturnableLinkableKeys extends string = string>(
+    groupIds: string[],
+    config?: RestoreReturn<TReturnableLinkableKeys>,
+    sharedContext?: Context
+  ): Promise<Record<TReturnableLinkableKeys, string[]> | void>
 }